chore&refactor: refactor and modify the project according to review comments on PR
