+++ cmd/cat/cat.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"bufio"
 	"flag"
 	"fmt"
 	"log"
-	"os"
+
+	"github.com/codescalersinternships/Coreutils-MohamedFadel/internal/cat"
 )
 
 func parseFlagsAndArguments() (bool, []string, error) {
@@ -23,39 +23,13 @@ func parseFlagsAndArguments() (bool, []string, error) {
 	return lineNumberingFlag, filesPaths, nil
 }
 
-func cat(lineNumberingFlag bool, filesPaths []string) error {
-	var lineCounter uint = 1
-
-	for _, path := range filesPaths {
-		file, err := os.Open(path)
-		if err != nil {
-			return fmt.Errorf("%s: cannot open this file for reading: No such file or directory", path)
-		}
-
-		scanner := bufio.NewScanner(file)
-
-		if lineNumberingFlag {
-			for scanner.Scan() {
-				fmt.Println(lineCounter, "	", scanner.Text())
-				lineCounter++
-			}
-		} else {
-			for scanner.Scan() {
-				fmt.Println(scanner.Text())
-			}
-		}
-	}
-
-	return nil
-}
-
 func main() {
 	lineNumberingFlag, filesPaths, err := parseFlagsAndArguments()
 	if err != nil {
 		log.Fatal(err)
 	}
-	err = cat(lineNumberingFlag, filesPaths)
-	if err != nil {
+
+	if err = cat.Cat(lineNumberingFlag, filesPaths); err != nil {
 		log.Fatal(err)
 	}
