+++ cmd/tree/tree.go
@@ -5,7 +5,8 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"strings"
+
+	"github.com/codescalersinternships/Coreutils-MohamedFadel/internal/tree"
 )
 
 func parseFlagsAndArguments() (uint, string, error) {
@@ -17,56 +18,16 @@ func parseFlagsAndArguments() (uint, string, error) {
 
 	path := flag.Arg(0)
 	if path == "" {
-		return 1, "", fmt.Errorf("invalid path, you have to specify a valid path to a directory")
+		currentDir, err := os.Getwd()
+		if err != nil {
+			return 1, "", fmt.Errorf("invalid path, you have to specify a valid path to a directory")
+		}
+		path = currentDir
 	}
 
 	return depthLevelFlag, path, nil
 }
 
-func tab(tabs uint) {
-	for i := 0; i < int(tabs); i++ {
-		fmt.Print("│   ")
-	}
-}
-
-func tree(path string, depthLevelFlag, numOfTabs uint, dirCount, fileCount *uint) error {
-	contents, err := os.ReadDir(path)
-	if err != nil {
-		return fmt.Errorf("%s: [error opening dir]", path)
-	}
-
-	for _, content := range contents {
-		if strings.HasPrefix(content.Name(), ".") {
-			continue
-		}
-
-		if content.IsDir() {
-			*dirCount++
-		} else {
-			*fileCount++
-		}
-
-		tab(numOfTabs)
-		fmt.Println("│──", content.Name())
-
-		if content.IsDir() && depthLevelFlag > 1 {
-			newPath := path + "/" + content.Name()
-			numOfTabs++
-			depthLevelFlag--
-
-			err := tree(newPath, depthLevelFlag, numOfTabs, dirCount, fileCount)
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			numOfTabs--
-			depthLevelFlag++
-		}
-
-	}
-
-	return nil
-}
 func main() {
 	depthLevelFlag, path, err := parseFlagsAndArguments()
 	if err != nil {
@@ -77,8 +38,7 @@ func main() {
 
 	var fileCount, dirCount uint = 0, 0
 
-	err = tree(path, depthLevelFlag, 0, &dirCount, &fileCount)
-	if err != nil {
+	if err := tree.Tree(path, depthLevelFlag, 0, &dirCount, &fileCount); err != nil {
 		log.Fatal(err)
 	}
