+++ internal/head/head.go
@@ -1,0 +1,25 @@
+package head
+
+import (
+	"bufio"
+	"fmt"
+	"os"
+)
+
+func Head(filePath string, numOfLines uint) error {
+	file, err := os.Open(filePath)
+	if err != nil {
+		return fmt.Errorf("cannot open this file for reading: No such file or directory")
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+
+	var lineCounter uint = 0
+	for scanner.Scan() && lineCounter < numOfLines {
+		line := scanner.Text()
+		lineCounter++
+		fmt.Println(line)
+	}
+	return nil
+}
