+++ internal/tail/tail.go
@@ -1,0 +1,33 @@
+package tail
+
+import (
+	"bufio"
+	"fmt"
+	"os"
+)
+
+func Tail(filePath string, numOfLines uint) error {
+	file, err := os.Open(filePath)
+	if err != nil {
+		return fmt.Errorf("cannot open this file for reading: No such file or directory")
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+
+	lines := make([]string, 0)
+
+	for scanner.Scan() {
+		lines = append(lines, scanner.Text())
+	}
+
+	if uint(len(lines)) > numOfLines {
+		lines = lines[len(lines)-int(numOfLines):]
+	}
+
+	for _, line := range lines {
+		fmt.Println(line)
+	}
+
+	return nil
+}
