+++ internal/yes/yes.go
@@ -1,0 +1,17 @@
+package yes
+
+import (
+	"fmt"
+	"strings"
+)
+
+func Yes(words []string) {
+	toBePrinted := "y"
+	if len(words) != 0 {
+		toBePrinted = strings.Join(words, " ")
+	}
+
+	for {
+		fmt.Println(toBePrinted)
+	}
+}
